# libc: Optimize ARM memcmp by using NEON.